+++ addons/godotdevconsole/scripts/Panel.cs
@@ -1,4 +1,3 @@
-
 using GodotDevConsole.Logging;
 using Godot;
 
@@ -6,9 +5,11 @@ namespace GodotDevConsole
 {
     public abstract partial class Panel : Control
     {
-        public virtual void SetPreActive(bool state) {}
-        public virtual void SetActive(bool state) {}
+        public virtual void PreActivate() { }
+        public virtual void PreDeactivate() { }
+        public virtual void Activate() { }
+        public virtual void Deactivate() { }
 
-        public virtual void Log(Log log, string logMessage) {}
+        public virtual void Log(Log log, string logMessage) { }
     }
 }