+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Replaced `SetPreActive` and `SetActive` with `PreActivate`, `PreDeactivate`, `Activate` and `Deactivate`.
 
 ## [1.3.0] - 2025-03-13